backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
tensorflow
numpy
python-multipart
google-generativeai
